Backend/requirements.txt
# # Core Framework
# fastapi==0.109.0
# uvicorn[standard]==0.27.0
# python-multipart==0.0.6

# # LlamaIndex
# llama-index==0.14.15
# llama-index-llms-gemini
# llama-index-embeddings-gemini
# llama-index-vector-stores-chroma

# # Vector Database
# chromadb==0.4.22

# # PDF Processing
# PyPDF2>=3.0.0
# pypdf>=4.0.0
# pdfplumber==0.10.3
# # EPUB
# ebooklib==0.18
# beautifulsoup4>=4.12.3,<5
# # DOCX
# python-docx==1.1.0
# # MOBI/AZW
# mobi==0.3.3
# lxml==5.1.0

# # Google AI (Gemini)
# google-generativeai==0.3.2

# # supabase auth
# supabase>=2.16.0

# # Utilities
# python-dotenv==1.0.0
# pydantic>=2.8,<3
# pydantic-settings>=2.3,<3

# # Testing
# pytest==7.4.4
# pytest-asyncio==0.23.3
# httpx==0.26.0

# python-jose[cryptography]>=3.3.0
# passlib[bcrypt]>=1.7.4
# aiofiles>=23.0.0

# pytest-cov>=4.1.0

# # Logging and Monitoring
# loguru>=0.6,<0.7



# Core Framework
fastapi==0.118.0
uvicorn[standard]==0.37.0
python-multipart==0.0.20

# LlamaIndex
llama-index==0.14.15
llama-index-llms-gemini==0.6.2
llama-index-embeddings-gemini==0.4.2
llama-index-vector-stores-chroma==0.5.5

# Vector Database
chromadb==1.5.1

# PDF and Documents
PyPDF2>=3.0.0
pypdf>=4.0.1
pdfplumber==0.10.3

ebooklib==0.18
beautifulsoup4>=4.12.3,<5
python-docx==1.1.0
mobi==0.3.3
lxml

# Google AI / Supabase
google-genai
google_api_core==2.27.0
supabase
python-jose[cryptography]>=3.3.0

# Utilities
python-dotenv==1.0.0
pydantic>=2.8,<3
pydantic-settings>=2.3,<3

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx>=0.27.0

python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
aiofiles>=23.0.0
pytest-cov>=4.1.0

# Logging
loguru>=0.6,<0.7
tenacity==8.2.3